requirements.txt
Flask==2.0.2
python-dotenv==0.19.0
requests==2.26.0
python-docx==0.8.11

